Update Auth.php for Custom Profile Saving
Added logic to allow custom profile fields to be saved back as part of the field mapping. TODO WARNING currently no data validation is done before saving custom fields back. Currently if the SAML data is not of the correct type save fails and an error is displayed to the user and they cannot log in!