Support for Moodle custom profile field mapping - part two. (#49)
* Updated Settings.html for Custom Profile Fields

Added logic to settings.html to enable mapping of saml claims to custom profile fields.

* Update Auth.php for Custom Profile Saving

Added logic to allow custom profile fields to be saved back as part of the field mapping. TODO WARNING currently no data validation is done before saving custom fields back. Currently if the SAML data is not of the correct type save fails and an error is displayed to the user and they cannot log in!

* Update Auth.php (Travis)

* Updates for Travis

* more updates for Travis

* Update auth.php

* Update settings.html

* Modification to method of obtaining custom user profile fields.