Issue #72, missing saml2 sitedata files when trying to login.
When refreshing a production database to a new development or staging server and the sitedata has not been copied over, the sitedata/saml2 directory will not be present.

During the loginpage_hook() function it will try to instantiate the saml2 auth and promptly error out. It could not read the idp.xml or associated certificate files.

Extra debugging has been added to function is_configured().

During loginpage_hook() it will now check for is_configured() and return if it is not setup correctly.