new: Block or redirect flagged accounts
Feature to allow Moodle admin to set custom IdP attribute and value to look for and redirect user or
display custom error message if the configured flag is passed by the IdP.

This includes supporting tests for parsing of the values and the redirect URL provided.

In order to set an attribute go to Site Administration > Plugins > Authentication > SAML2

Under the 'Flagged account blocking / redirection' section, set the attribute and value you wish to look
for and the response type you wish to use (redirect to a URL or display a custom message)

You then need to use your IdP to set a custom flag string and value string on flagged accounts.