improvement: Added additional validation to handle_flag_login
Added additional validation when handling IdP attributes in order to avoid allowing empty string attributes to lead to a redirect.
Empty string attribute values should be allowed, as a user may choose to look for an empty string value for an attribute as the flag, but the attribute itself cannot be an empty string.