Restrict automatic account creation to users in "allowed email domains"
Moodle has a core setting -- "allowemailaddresses" -- that you can use
to restrict access to email accounts in particular domains. It is used
by the "Email-based self-registration" and "OAuth2" authentication
plugins.

This change allows automatic account creation for users whose mail
attribute matches "allowemailaddresses".  But it does not prevent
you from creating a Moodle account for a user who has an address in
a different domain, and is authenticated by your IdP.

Fixes #412