Add support for a mock IdP for use in Behat tests
Adds a minimal mock IdP and simple Behat steps which can be used to
easily test aspects of this plugin, or to accurately simulate a SAML
login as part of tests elsewhere in a system.

Replaces the existing fake-IdP support, which used SimpleSamlPHP and
required a separate hostname (and was generally more complicated).