Update Travis to use latest Node version, now required by Moodle tools
Note that 'nvm install' with no number doesn't work because there is
no .nvmrc from Moodle available at that point. So I added the required
versions (which are different for MOODLE_36_STABLE as it is out of
support) to an environment variable.

Ideally, there would be a change to moodle-plugin-ci to do 'nvm use'
after installing Moodle, which should make this work permanently,
but this change will fix it until then.

Also fixes a missing NL at EOF in a test file which causes a failure
in master now.