Fix auto-login when going directly to activity
With Moodle autologinguest and SAML autologin both on, if you go to
an activity when you are not logged in, you get redirected to
enrol/index.php for the course before the SAML autologin feature
triggers. Autologin does happen correctly on the enrol page, but
after a successful autologin (if you now have access to the course)
you end up on the course page, not the activity page.

This change fixes that behaviour so that after successful autologin
you get to the activity page that you originally requested, using
$SESSION->wantsurl (which is set by require_login just before it
redirects you to the enrol page).