Fix warning on logout.
Before registering a logout handler check that session is still
authenticated. If logout initiated from Moodle, the session will be
destroyed earlier, so no need to register a handler (we logged out user
in Moodle already anyway). When logout is initiated from IdP, the SP
session will be active, so we register handler that will perform logout
in Moodle when session is destroyed by SP.

Closes #520