auth/saml2 Fully hydrate user account before creating it
This prevents issues with plugins such as local welcome, where they
cannot function correctly as the account is created with only minimal
information such as the username.

This extracts the logic for extracting saml2 attributes out of the
update function so both update and create can use it, and then fully
hydrates a user object before calling user_create_user to create the
moodle user object.

Password updating is set to true for the use case where a different
auth plugin is being used that might support the password field

Custom profile fields are saved just after user creation

This ensures the record has any custom profile fields filled out,
although note that they will not appear in the create user event as that
happens first.

There is also a fix for the unit test for user creation

Now that user creation is fully hydrating, there is not a user update
event sent, as the code only triggers a user update if fields have
changed.

This means a user creation on login only emits two events, the user
created event and then the login event.