Fix bug where user account could not be found when user has multiple accounts with same email address
This fixes a bug that occured when a users tried to login using SAML2. The conditions are:

- User has multiple Moodle accounts with the same email address (or other mapped idP field attributes that are being used to find and login the user in Moodle).

The condition was count($records == 1) so a user was being found only when there was exactly 1 user account found. This would make the hole reset() call on the next line redundant because there would always be exactly 0 or 1 records...

So I think this needs to be count($records) > 0 so every scenario where the sql query returns multiple users from the Moodle database, the forst one will always be used.